# perf(encoding): Comparing row and data_chunk benchmarks